got my own shader to work

+++ MineDirt/Src/Blocks.cs
@@ -22,9 +22,9 @@ public static class Blocks
     // [0] = Front, Back, Left, Right, [1] = Top, Bottom
     // [0] = Front, Back, Left, Right, [1] = Top, [2] = Bottom
     // [0] = Front, [1] = Back, [2] = Left, [3] = Right, [4] = Top, [5] = Bottom]
-    public static Block Dirt(Vector3 pos) => new(BlockType.Dirt, [2]);
-    public static Block Grass(Vector3 pos) => new(BlockType.Grass, [1, 0, 2]);
-    public static Block Cobblestone(Vector3 pos) => new(BlockType.Cobblestone, [3]);
-    public static Block Bedrock(Vector3 pos) => new(BlockType.Bedrock, [4]);
-    public static Block Stone(Vector3 pos) => new(BlockType.Stone, [5]);
+    public static Block Dirt(Vector3 pos) => new((byte)BlockType.Dirt, [2]);
+    public static Block Grass(Vector3 pos) => new((byte)BlockType.Grass, [1, 0, 2]);
+    public static Block Cobblestone(Vector3 pos) => new((byte)BlockType.Cobblestone, [3]);
+    public static Block Bedrock(Vector3 pos) => new((byte)BlockType.Bedrock, [4]);
+    public static Block Stone(Vector3 pos) => new((byte)BlockType.Stone, [5]);
 }
